move things to server
