--- back/polkassembly-chain-db-app-server/.gitignore
@@ -1,7 +1,0 @@
-.env*
-dist
-package-lock.json
-node_modules
-.idea
-.vscode
-*.log
