--- back/polkassembly-chain-db-app-server/Dockerfile
@@ -1,8 +1,0 @@
-FROM node:alpine3.10
-
-WORKDIR /server
-COPY . .
-
-RUN yarn
-
-CMD ["yarn", "start"]
