--- back/polkassembly-chain-db-app-server/kubernetes/loadbalancer.yaml
@@ -1,17 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: polkassembly-chain-db-app-server
-  namespace: polkassembly-chain-db-app-server
-  labels:
-    stage: staging
-    name: polkassembly-chain-db-app-server
-    app: polkassembly-chain-db-app-server
-spec:
-  type: LoadBalancer
-  selector:
-    app: polkassembly-chain-db-app-server
-  ports:
-    - protocol: TCP
-      port: 4010
-      targetPort: 4010
