--- back/polkassembly-chain-db-app-server/kubernetes/server-deployment.yaml
@@ -1,32 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: polkassembly-chain-db-app-server
-  namespace: polkassembly-chain-db-app-server
-  labels:
-    stage: staging
-    name: polkassembly-chain-db-app-server
-    app: polkassembly-chain-db-app-server
-spec:
-  selector:
-    matchLabels:
-      app: polkassembly-chain-db-app-server
-  replicas: 1
-  strategy:
-    type: Recreate
-  template:
-    metadata:
-      labels:
-        stage: staging
-        name: polkassembly-chain-db-app-server
-        app: polkassembly-chain-db-app-server
-    spec:
-      containers:
-        - name: polkassembly-chain-db-app-server
-          image: eu.gcr.io/test-installations-222013/nomidot_server@sha256:8f7cb5a9eeff909431cd195b606f9a662c289dcd4d842e2c86555906a03364d5
-          env:
-            - name: PRISMA_ENDPOINT
-              value: http://10.0.9.43:4466
-          ports:
-            - name: polkassembly-chain-db-app-server-4010
-              containerPort: 4010
