--- back/polkassembly-chain-db-app-server/prisma/datamodel.prisma
@@ -1,159 +1,0 @@
-
-type BlockNumber {
-  id: ID! @id
-  number: Int! @unique
-  authoredBy: String!
-  startDateTime: DateTime!
-  hash: String! @unique
-}
-
-type Session {
-  id: ID! @id
-  index: Int! @unique
-  start: BlockNumber! @relation(name: "BlockNumberAtSessionStart", link: INLINE)
-}
-
-type Era {
-  id: ID! @id
-  index: Int! @unique
-  eraStartSessionIndex: Session! @relation(name: "EraStartSessionIndex", link: INLINE)
-  totalPoints: String!
-  individualPoints: [String!]! @scalarList(strategy: RELATION)
-}
-
-type Slashing {
-  id: ID! @id
-  blockNumber: BlockNumber! @relation(name: "SlashingAtBlockNumber", link: INLINE)
-  who: String! @unique
-  amount: String!
-}
-
-type Validator {
-  id: ID! @id
-  session: Session! @relation(name: "ValidatorsAtSession", link: INLINE)
-  controller: String!
-  stash: String!
-  preferences: String!
-}
-
-type TotalIssuance {
-  id: ID! @id
-  blockNumber: BlockNumber! @relation(name: "TotalIssuanceAtBlockNumber", link: INLINE)
-  amount: String!
-}
-
-type Reward {
-  id: ID! @id
-  amount: String!
-  authoredBlock: BlockNumber! @relation(name: "RewardsAtBlockNumber", link: INLINE)
-  recipients: [String!]! @scalarList(strategy: RELATION)
-}
-
-type Nomination {
-  id: ID! @id
-  validatorController: String!
-  validatorStash: String!
-  nominatorController: String!
-  nominatorStash: String!
-  session: Session! @relation(name: "NominationsAtSession", link: INLINE)
-  bonded: String!
-}
-
-type Stake {
-  id: ID! @id
-  blockNumber: BlockNumber! @relation(name: "StakeAtBlockNumber", link: INLINE)
-  totalStake: String!
-}
-
-type Motion {
-  id: Int! @id(strategy: SEQUENCE) @sequence(name: "MOTION_SEQUENCE" initialValue:1 allocationSize:1)
-  author: String!
-  memberCount: Int!
-  metaDescription: String!
-  method: String!
-  motionProposalArguments: [MotionProposalArgument]! @relation(name: "ArgumentForMotionProposal")
-  motionProposalHash: String!
-  motionProposalId: Int! @unique
-  preimage: Preimage @relation(name: "PreimageForMotion")
-  preimageHash: String
-  section: String!
-  status: [MotionStatus]! @relation(name: "MotionStatusFromMotion")
-}
-
-type MotionProposalArgument {
-  id: ID! @id
-  name: String!
-  motion: Motion! @relation(name: "ArgumentForMotionProposal", link: INLINE)
-  value: String!
-}
-
-type MotionStatus {
-  id: ID! @id
-  blockNumber: BlockNumber! @relation(name: "MotionStatusAtBlockNumber", link: INLINE)
-  motion: Motion! @relation(name: "MotionStatusFromMotion", link: INLINE)
-  status: String!
-}
-
-type Proposal {
-  id: Int! @id(strategy: SEQUENCE) @sequence(name: "PROPOSAL_SEQUENCE" initialValue:1 allocationSize:1)
-  author: String!
-  depositAmount: String!
-  preimage: Preimage @relation(name: "PreimageForProposal")
-  preimageHash: String!
-  proposalId: Int! @unique
-  proposalStatus: [ProposalStatus]! @relation(name: "ProposalStatusFromProposal")
-}
-
-type ProposalStatus {
-  id: ID! @id
-  blockNumber: BlockNumber! @relation(name: "ProposalStatusAtBlockNumber", link: INLINE)
-  proposal: Proposal! @relation(name: "ProposalStatusFromProposal", link: INLINE)
-  status: String!
-}
-
-type Preimage {
-  id: ID! @id
-  author: String!
-  depositAmount: String!
-  hash: String!
-  metaDescription: String!
-  method: String!
-  motion: Motion @relation(name: "PreimageForMotion", link: INLINE)
-  proposal: Proposal @relation(name: "PreimageForProposal", link: INLINE)
-  preimageArguments: [PreimageArgument]! @relation(name: "ArgumentForPreimage")
-  preimageStatus: [PreimageStatus]! @relation(name: "PreimageStatusFromPreimage")
-  referendum: Referendum @relation(name: "PreimageForReferendum", link: INLINE)
-  section: String!
-}
-
-type PreimageStatus {
-  id: ID! @id
-  blockNumber: BlockNumber! @relation(name: "PreimagStatusAtBlockNumber", link: INLINE)
-  preimage: Preimage! @relation(name: "PreimageStatusFromPreimage", link: INLINE)
-  status: String!
-}
-
-type PreimageArgument {
-  id: ID! @id
-  name: String!
-  preimage: Preimage! @relation(name: "ArgumentForPreimage", link: INLINE)
-  value: String!
-}
-
-type Referendum {
-  id: Int! @id(strategy: SEQUENCE) @sequence(name: "REFERENDUM_SEQUENCE" initialValue:1 allocationSize:1)
-  delay: Int!
-  end: Int!
-  preimage: Preimage @relation(name: "PreimageForReferendum")
-  preimageHash: String!
-  referendumId: Int! @unique
-  referendumStatus: [ReferendumStatus]! @relation(name: "ReferendumStatusFromReferendum")
-  voteThreshold: String!
-}
-
-type ReferendumStatus {
-  id: ID! @id
-  blockNumber: BlockNumber! @relation(name: "ReferendumStatusAtBlockNumber", link: INLINE)
-  referendum: Referendum! @relation(name: "ReferendumStatusFromReferendum", link: INLINE)
-  status: String!
-}
