--- back/polkassembly-chain-db-app-server/prisma/prisma.yml
@@ -1,19 +1,0 @@
-# Specifies the HTTP endpoint of your Prisma API (deployed to a Prisma Demo server).
-endpoint: ${env:PRISMA_ENDPOINT}
-
-# Defines your models, each model is mapped to the database as a table.
-datamodel: datamodel.prisma
-
-# Specifies the language and directory for the generated Prisma client.
-generate:
-  - generator: typescript-client
-    output: ../src/generated/prisma-client
-
-# Seed your service with initial data based on `seed.graphql`.
-# seed:
-#   import: seed.graphql
-
-# Ensures Prisma client is re-generated after a datamodel change.
-hooks:
-  post-deploy:
-    - yarn prisma generate
