--- back/polkassembly-chain-db-app-server/src/schema.graphql
@@ -1,32 +1,0 @@
-
-# import * from './generated/prisma.graphql'
-
-type Query {
-  blockNumbers(where: BlockNumberWhereInput, orderBy: BlockNumberOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [BlockNumber]!
-  preimages(where: PreimageWhereInput, orderBy: PreimageOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [Preimage]!
-  preimageArguments(where: PreimageArgumentWhereInput, orderBy: PreimageArgumentOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [PreimageArgument]!
-  preimageStatuses(where: PreimageStatusWhereInput, orderBy: PreimageStatusOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [PreimageStatus]!
-  proposals(where: ProposalWhereInput, orderBy: ProposalOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [Proposal]!
-  proposalStatuses(where: ProposalStatusWhereInput, orderBy: ProposalStatusOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [ProposalStatus]!
-  referendums(where: ReferendumWhereInput, orderBy: ReferendumOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [Referendum]!
-  referendumStatuses(where: ReferendumStatusWhereInput, orderBy: ReferendumStatusOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [ReferendumStatus]!
-  motion(where: MotionWhereUniqueInput!): Motion
-  motions(where: MotionWhereInput, orderBy: MotionOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [Motion]!
-  motionProposalArgument(where: MotionProposalArgumentWhereUniqueInput!): MotionProposalArgument
-  motionProposalArguments(where: MotionProposalArgumentWhereInput, orderBy: MotionProposalArgumentOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [MotionProposalArgument]!
-  motionStatus(where: MotionStatusWhereUniqueInput!): MotionStatus
-  motionStatuses(where: MotionStatusWhereInput, orderBy: MotionStatusOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [MotionStatus]!
-  blockNumber(where: BlockNumberWhereUniqueInput!): BlockNumber
-  preimage(where: PreimageWhereUniqueInput!): Preimage
-  preimageArgument(where: PreimageArgumentWhereUniqueInput!): PreimageArgument
-  preimageStatus(where: PreimageStatusWhereUniqueInput!): PreimageStatus
-  proposal(where: ProposalWhereUniqueInput!): Proposal
-  proposalStatus(where: ProposalStatusWhereUniqueInput!): ProposalStatus
-  referendum(where: ReferendumWhereUniqueInput!): Referendum
-  referendumStatus(where: ReferendumStatusWhereUniqueInput!): ReferendumStatus
-}
-
-type Subscription {
-  subscribeProposal: Proposal
-  subscribeReferendum: Referendum
-}
