--- back/polkassembly-chain-db-app-server/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "moduleResolution": "node",
-    "module": "commonjs",
-    "sourceMap": true,
-    "rootDir": "src",
-    "outDir": "dist",
-    "lib": [
-      "esnext", "dom"
-    ]
-  }
-}
