+++ back/server/src/schema.graphql
@@ -24,7 +24,6 @@ type Query {
   
   validators(where: ValidatorWhereInput, orderBy: ValidatorOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [Validator]!
 
-  blockNumber(where: BlockNumberWhereUniqueInput!): BlockNumber
   era(where: EraWhereUniqueInput!): Era
   nomination(where: NominationWhereUniqueInput!): Nomination
   reward(where: RewardWhereUniqueInput!): Reward
@@ -33,6 +32,27 @@ type Query {
   stake(where: StakeWhereUniqueInput!): Stake
   totalIssuance(where: TotalIssuanceWhereUniqueInput!): TotalIssuance
   validator(where: ValidatorWhereUniqueInput!): Validator
+  preimages(where: PreimageWhereInput, orderBy: PreimageOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [Preimage]!
+  preimageArguments(where: PreimageArgumentWhereInput, orderBy: PreimageArgumentOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [PreimageArgument]!
+  preimageStatuses(where: PreimageStatusWhereInput, orderBy: PreimageStatusOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [PreimageStatus]!
+  proposals(where: ProposalWhereInput, orderBy: ProposalOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [Proposal]!
+  proposalStatuses(where: ProposalStatusWhereInput, orderBy: ProposalStatusOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [ProposalStatus]!
+  referendums(where: ReferendumWhereInput, orderBy: ReferendumOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [Referendum]!
+  referendumStatuses(where: ReferendumStatusWhereInput, orderBy: ReferendumStatusOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [ReferendumStatus]!
+  motion(where: MotionWhereUniqueInput!): Motion
+  motions(where: MotionWhereInput, orderBy: MotionOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [Motion]!
+  motionProposalArgument(where: MotionProposalArgumentWhereUniqueInput!): MotionProposalArgument
+  motionProposalArguments(where: MotionProposalArgumentWhereInput, orderBy: MotionProposalArgumentOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [MotionProposalArgument]!
+  motionStatus(where: MotionStatusWhereUniqueInput!): MotionStatus
+  motionStatuses(where: MotionStatusWhereInput, orderBy: MotionStatusOrderByInput, skip: Int, after: String, before: String, first: Int, last: Int): [MotionStatus]!
+  blockNumber(where: BlockNumberWhereUniqueInput!): BlockNumber
+  preimage(where: PreimageWhereUniqueInput!): Preimage
+  preimageArgument(where: PreimageArgumentWhereUniqueInput!): PreimageArgument
+  preimageStatus(where: PreimageStatusWhereUniqueInput!): PreimageStatus
+  proposal(where: ProposalWhereUniqueInput!): Proposal
+  proposalStatus(where: ProposalStatusWhereUniqueInput!): ProposalStatus
+  referendum(where: ReferendumWhereUniqueInput!): Referendum
+  referendumStatus(where: ReferendumStatusWhereUniqueInput!): ReferendumStatus
 }
 
 type Subscription {
@@ -46,4 +66,7 @@ type Subscription {
   subscribeSlashings: Slashing
   subscribeStakes: Stake
   subscribeValidators: Validator
+  subscribeMotion: Motion
+  subscribeProposal: Proposal
+  subscribeReferendum: Referendum
 }
